Make comment more objective
The way that different packages handle requirements files is based on the personal preference of the package author. Some authors feel that it is bad practice to read from requirements files in the setup.py file, because requirements files are typically used to pin dependencies to particular versions. Other authors feel that the opposite is bad practice. It's a difference of opinion. :sunglasses: