Refactor packagecloud integration
The cookiecutter will always ask for a packagecloud repo and master
token. This is a safe assumption (since the cookiecutter is only used
within Syapse) and it's low-risk even if they don't pull anything from
packagecloud. Having the additional source available will lower the
barrier to entry for using other Syapse packages later.

The logic for configuring packagecloud authentication was moved into a
`bin/setup` script. This ensures that we don't break `cookiecutter`'s
`--no-input` and `--replay` options.

The `bin/setup` command essentially writes to a `.env` file which is
sourced by `pipenv` when a developer wants to work on the project. The
script will prompt them for a token if `PACKAGECLOUD_TOKEN` isn't set.